+++ CHANGELOG.md
@@ -13,6 +13,10 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 
 ## v3.?.? (2021-??-??)
 
+#### :boom: Breaking Change
+
+* The module has been moved to `promise/abortable` `core/then`
+
 #### :rocket: New Feature
 
 * Added a new static method `allSettled`: